requirements.txt
torch>=2.0.0
pytorch-lightning>=2.0.0
pytorch-forecasting>=1.0.0
hmmlearn>=0.3.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
tensorboard>=2.14.0
openpyxl>=3.1.0
xlrd>=2.0.0
scipy>=1.10.0
pytest>=7.0.0
